Move executor package to runtime
